extract parser module
