+++ extensions/positron-code-cells/src/codeLenses.ts
@@ -3,46 +3,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as vscode from 'vscode';
-
-
-export interface ICell {
-	range: vscode.Range;
-}
-
-
-class CellMatcher {
-	isCell(line: string): boolean {
-		return line.startsWith('# %%');
-	}
-}
-
-
-export function generateCellRangesFromDocument(document: vscode.TextDocument): ICell[] {
-	// Implmentation of getCells here based on Don's Jupyter extension work
-	const matcher = new CellMatcher();
-	const cells: ICell[] = [];
-	for (let index = 0; index < document.lineCount; index += 1) {
-		const line = document.lineAt(index);
-		if (matcher.isCell(line.text)) {
-			if (cells.length > 0) {
-				const previousCell = cells[cells.length - 1];
-				previousCell.range = new vscode.Range(previousCell.range.start, document.lineAt(index - 1).range.end);
-			}
-
-			cells.push({
-				range: line.range,
-			});
-		}
-	}
-
-	if (cells.length >= 1) {
-		const line = document.lineAt(document.lineCount - 1);
-		const previousCell = cells[cells.length - 1];
-		previousCell.range = new vscode.Range(previousCell.range.start, line.range.end);
-	}
-
-	return cells;
-}
+import { parseCells } from './parser';
 
 function runCellCodeLens(range: vscode.Range, line: number): vscode.CodeLens {
 	return new vscode.CodeLens(range, {
@@ -80,7 +41,7 @@ export function registerCodeLensProvider(context: vscode.ExtensionContext): void
 				}
 
 				const codeLenses: vscode.CodeLens[] = [];
-				const cells = generateCellRangesFromDocument(document);
+				const cells = parseCells(document);
 				for (let i = 0; i < cells.length; i += 1) {
 					const cell = cells[i];
 					const range = cell.range;
