+++ extensions/positron-code-cells/src/folding.ts
@@ -2,13 +2,13 @@
  *  Copyright (C) 2023 Posit Software, PBC. All rights reserved.
  *--------------------------------------------------------------------------------------------*/
 import * as vscode from 'vscode';
-import { generateCellRangesFromDocument } from './codeLenses';
+import { parseCells } from './parser';
 
 export function registerFoldingRangeProvider(context: vscode.ExtensionContext): void {
 	context.subscriptions.push(
 		vscode.languages.registerFoldingRangeProvider('*', {
 			provideFoldingRanges: (document) =>
-				generateCellRangesFromDocument(document).map((cell) =>
+				parseCells(document).map((cell) =>
 					new vscode.FoldingRange(cell.range.start.line, cell.range.end.line)
 				)
 		}),
